+++ sr_unix/ydb_timer_cancel.c
@@ -1,6 +1,6 @@
 /****************************************************************
  *								*
- * Copyright (c) 2018-2019 YottaDB LLC and/or its subsidiaries.	*
+ * Copyright (c) 2018-2020 YottaDB LLC and/or its subsidiaries.	*
  * All rights reserved.						*
  *								*
  *	This source code contains the intellectual property	*
@@ -19,7 +19,7 @@
 #include "libydberrors.h"
 
 /* Simple YottaDB wrapper for gtm_cancel_timer() */
-void ydb_timer_cancel(int timer_id)
+void ydb_timer_cancel(intptr_t timer_id)
 {
 	boolean_t	error_encountered;
 	DCL_THREADGBL_ACCESS;
